Continue NewUserForm build
Design pattern is to pull out at much logic from the component
for ease with later testing; create test todos as needed to
allow for continued development.

Note also that the three input fields are now wrapped
in labels; this allows for easier querying of the elements
for testing and improved accessibility.