Add routing and navigation tests
TIL: MemoryRouter is the proper choice for testing, as the
clicks (which trigger navigate) trigger the desired re-render.
Router works for more static tests, like testing
that the routes defined in App work as intended.

This commit adds routes for Landing, Profile, Register, and Login,
tests those routes, and tests that registering or login trigger
navigation to Profile as intended.